Move citation comp to cite folder
